requirements.txt
fmpsdk
python-dotenv
matplotlib
numpy
